Create Cars
The store that you're currently viewing is based off cars that i want or have already driven; with that being said, i started off by creating the database for the store by carefully choosing columns such as pricing, the year the car was made and number of cars sold (based off numbers i pulled out of thin air). Following up with giving a sum of the all cars sold. And just a quick disclaimer the prices and overall numbers in my store are based solely off my desires; meaning the prices displayed are more than likely not true but what I thought they are.