Create Sneakers
The store that you're currently viewing is based on sneakers that I have in my possession or currently in pursuit of; with that being said, I started off by creating the database for the store by carefully choosing columns such as pricing, choosing at random the sizing, number of shoes sold (based off numbers I pulled out of thin air). Following up with giving a sum of the all sneakers sold. And just a quick disclaimer the prices and overall numbers in my store are based solely off my desires; meaning the prices displayed are more than likely not true but what I thought they are.